requirements.txt
# Core
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0

# PDF Processing
pypdf>=4.0.0
pdfplumber>=0.11.0

# Vector Store
faiss-cpu>=1.7.4
chromadb>=0.5.0

# Reranking (cross-encoder, no extra API key needed)
sentence-transformers>=3.0.0

# Web Search / Scraping
duckduckgo-search>=6.2.0    # No API key needed
beautifulsoup4>=4.12.0
requests>=2.31.0

# API Server
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9     # for file upload

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.7.0